+++ podoc/plugins/ast/__init__.py
@@ -8,4 +8,4 @@
 # Imports
 #------------------------------------------------------------------------------
 
-from ._json import JSON
+from ._ast import AST
