+++ podoc/plugins/ast/_ast.py
@@ -17,6 +17,17 @@ logger = logging.getLogger(__name__)
 
 
 #------------------------------------------------------------------------------
+# Utils
+#------------------------------------------------------------------------------
+
+def ae(a, b):
+    if isinstance(a, (list, dict)):
+        assert _remove_json_meta(a) == _remove_json_meta(b)
+    else:
+        assert a == b
+
+
+#------------------------------------------------------------------------------
 # AST
 #------------------------------------------------------------------------------
