--- podoc/plugins/json/tests/test_json.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Test JSON plugin."""
-
-
-#------------------------------------------------------------------------------
-# Imports
-#------------------------------------------------------------------------------
-
-
-#------------------------------------------------------------------------------
-# Test JSON plugin
-#------------------------------------------------------------------------------
