+++ podoc/plugins/markdown/__init__.py
@@ -7,5 +7,3 @@
 #------------------------------------------------------------------------------
 # Imports
 #------------------------------------------------------------------------------
-
-from ._markdown import Markdown
