--- podoc/tests/test_testing.py
@@ -1,29 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Test testing functions."""
-
-
-#------------------------------------------------------------------------------
-# Imports
-#------------------------------------------------------------------------------
-
-from pytest import raises
-
-from ..testing import ae
-
-
-#------------------------------------------------------------------------------
-# Tests
-#------------------------------------------------------------------------------
-
-def test_ae():
-    ae(1, 1)
-    ae(1., 1)
-
-    ae({'a': 1, 'm': {}}, {'a': 1})
-    with raises(AssertionError):
-        ae({'a': 1, 'b': {}}, {'a': 1})
-
-    ae('abc\n', 'abc\n')
-    with raises(AssertionError):
-        ae('abc\n', 'abc')
